Simplify extend function
Instead of creatin a temporary array of all they keys of the ‘add’
object, simply iterate over them using for-in construct instead.
This makes the code shorter and, as hinted above, gets rid of an
unnecessary array.  The patch also changes the function to have
single return which is also nice.