improve performance somewhat
Quite a lot, actually.

Still a pathological case where a '**' pattern is followed by multiple
glob patterns, especially if the subsequent set of patterns includes
another globstar with the same problem, it seems to grow geometrically.

So '**/*/*.txt' takes about twice as long as it should, and
'**/*/**/*/*.txt' takes about twice as long again, increasining
geometrically.

Need to rethink the globstar handling approach, maybe the whole shape of
the walker class.  Currently, it's pretty naive, and the proof of
concept I added to test my assumption only works if there'e exactly 1
path portion following the globstar.  That is, it makes **/*.txt go as
fast as it should, but **/*/*.txt is still slow.

Right now, a test for **/*/*.txt in a path a/b/c, with directory child
paths [d, e] will create child walkers for the following paths/patterns:

*/*.txt a/b/c      Test without the ** portion, against this path
**/*/*.txt a/b/c/d Test with the globstar in each child directory
**/*/*.txt a/b/c/e
*/*.txt a/b/c/d    Test without globstar in each child directory
*/*.txt a/b/c/e

This happens at each directory level, so if d and e both have directory
children x and y, you get:

*/*.txt a/b/c/d
**/*/*.txt a/b/c/d/x
**/*/*.txt a/b/c/d/y
*/*.txt a/b/c/d/x
*/*.txt a/b/c/d/y
*/*.txt a/b/c/e
**/*/*.txt a/b/c/e/x
**/*/*.txt a/b/c/e/y
*/*.txt a/b/c/e/x
*/*.txt a/b/c/e/y

Added a check when there's exactly 1 entry after the globstar, but this
actually has a bug and isn't scalable anyway.