avoid walking same path/pattern repeatedly
Still a bit slower than I'd like, but no longer subject to _quite_ as
pathological geometric explosions on patterns with multiple ** portions.

I think at this point, the way to squeeze out the last bit of
performance is just cutting out a lot of the string and array
operations.  They're convenient and make the code a bit more elegant and
readable, but also it seems to be where it's spending most of the time.

Rather than doing a lot of 'const [p, ...rest] = pattern' types of code,
and then passing a pattern array to each child walker, what we could do
is pass the _same_ pattern array to every child in the walk, along with
an index telling it where to start.

This will be a bit of an annoying refactor, and make the code a bit less
elegant, but it should yield a considerable performance increase.  Along
with that, it'll have to pass that index into the stringPattern()
method, or maybe just pass a patternString index as well as a pattern
index, indicating how far in the single shared patternString it should
operate on.

Alternatively, it could pass the globParts array along with pattern
array, which would make this.patternString a simple array subset string
concat op.

Once that's done, the other potential idea to try to get even faster is
to use worker threads for readdir in the async case.  Not sure if
that'll be beneficial in the end, since node async fs ops already fan
out to multiple threads, or if it might have other undesirable knock-on
effects from using ALL THE CPUs for glob activities, but it could be an
interesing approach.

Note on the benchmarks below, since globby has essentially equivalent
performance to all the other fastglob-based implementations, and has the
closest API to node-glob, I don't include any of the others in the
benchmark script, since they're a bit more of a pain to do an
apples-to-apples comparison with, and in the end didn't add much
information.  Using fastglob directly _is_ slightly faster than globby,
but not by enough to really matter.

Benchmarks with this change:

    $ npm run bench

    > glob@9.0.0-0 bench
    > bash benchmark.sh

    --- pattern: './**/?/**/?/**/?/**/?/**/*.txt' ---
    node globby sync              	0m0.423s
    node globby async             	0m0.300s
    node current glob.sync cjs    	0m0.702s
    node current glob async cjs   	0m0.621s
    node current glob.sync mjs    	0m0.701s
    node current glob async mjs   	0m0.639s

^ still taking about twice as much time as it should, but no longer
increasing geometrically with each ** portion.

    --- pattern: './**/*/**/*/**/*/**/*/**/*.txt' ---
    node globby sync              	0m0.412s
    node globby async             	0m0.309s
    node current glob.sync cjs    	0m0.718s
    node current glob async cjs   	0m0.641s
    node current glob.sync mjs    	0m0.736s
    node current glob async mjs   	0m0.657s

^ same here, these two were really bad before, like 1-2 seconds

    --- pattern: './**/0/**/0/**/0/**/0/**/*.txt' ---
    node globby sync              	0m0.297s
    node globby async             	0m0.189s
    node current glob.sync cjs    	0m0.411s
    node current glob async cjs   	0m0.316s
    node current glob.sync mjs    	0m0.408s
    node current glob async mjs   	0m0.345s

    --- pattern: './**/[01]/**/[12]/**/[23]/**/[45]/**/*.txt' ---
    node globby sync              	0m0.297s
    node globby async             	0m0.185s
    node current glob.sync cjs    	0m0.416s
    node current glob async cjs   	0m0.281s
    node current glob.sync mjs    	0m0.421s
    node current glob async mjs   	0m0.282s

    --- pattern: './**/0/**/0/**/*.txt' ---
    node globby sync              	0m0.302s
    node globby async             	0m0.193s
    node current glob.sync cjs    	0m0.392s
    node current glob async cjs   	0m0.298s
    node current glob.sync mjs    	0m0.405s
    node current glob async mjs   	0m0.329s

^ getting SOOO CLOSE to fastglob/globby!

    --- pattern: '**/*.txt' ---
    node globby sync              	0m0.410s
    node globby async             	0m0.296s
    node current glob.sync cjs    	0m0.402s
    node current glob async cjs   	0m0.253s
    node current glob.sync mjs    	0m0.415s
    node current glob async mjs   	0m0.259s

^ basically exactly where it should be, but faster = better

    --- pattern: './**/**/**/**/**/**/**/**/*.txt' ---
    node globby sync              	0m0.408s
    node globby async             	0m0.292s
    node current glob.sync cjs    	0m0.399s
    node current glob async cjs   	0m0.257s
    node current glob.sync mjs    	0m0.412s
    node current glob async mjs   	0m0.290s

    --- pattern: '**/*/*.txt' ---
    node globby sync              	0m0.412s
    node globby async             	0m0.305s
    node current glob.sync cjs    	0m0.433s
    node current glob async cjs   	0m0.300s
    node current glob.sync mjs    	0m0.442s
    node current glob async mjs   	0m0.321s

    --- pattern: '**/*/**/*.txt' ---
    node globby sync              	0m0.416s
    node globby async             	0m0.300s
    node current glob.sync cjs    	0m0.476s
    node current glob async cjs   	0m0.337s
    node current glob.sync mjs    	0m0.480s
    node current glob async mjs   	0m0.346s

^ perf issue almost completely gone

    --- pattern: '**/[0-9]/**/*.txt' ---
    node globby sync              	0m0.401s
    node globby async             	0m0.294s
    node current glob.sync cjs    	0m0.478s
    node current glob async cjs   	0m0.334s
    node current glob.sync mjs    	0m0.494s
    node current glob async mjs   	0m0.346s

^ same here