more hacking to try to squeeze out some excess path.resolve() calls
Really, what has to happen here is instead of just relying on getting a
string, take over the path resolution by making the walker intelligently
track what its actual start directory will be.

When we pull string portions off the path, calculate the start directory
based on the path portions being pulled off.  Ie, if we see .., go up a
dir, if we see '' or '.', then leave it as is, anything else append with
'/'.

The result should be that every GlobWalker always has the
posix-separator-style absolute path, _without_ having to call
path.resolve() except at the top level if the pattern and cwd are both
not absolute.

Also, need tests for the Pattern class, that's why that's failing
coverage now.