new walking methodology, much faster for most patterns
This is still showing some pathological performance in the case
where several '..' and '**' pattern portions are used together,
for example with './*/**/../*/**/../*/**/../*/**/../*/**/*.txt'

In all other benchmark cases, it's now out-performing fast-glob/globby,
so I think this is the approach that'll stick!  Just need to figure out
how to handle those parent walks a bit more cleverly.