some pattern editing hacks (belongs in minimatch)
Identified a bunch of ways that patterns can be altered in
behavior-identical ways, to minimize the number of full-tree walks that
are required.  There's a slight perf hit up front with some of them,
like the long repeated string of '*/**/..', because it's doing a lot of
calculations that really ought to be done up front in the Minimatch
constructor, rather than here in the Pattern object.

If done *before* parsing the glob patterns into regexps and such, it'll
be much more efficient, but it took a bit of fiddling to get the correct
outcomes, so now the next step is porting it into Minimatch properly,
along with a ton of tests.

Also, this improves performance considerably by making the GlobUtil
classes able to take a set of Pattern objects, so that they can track
what patterns have already been walked, even when they appear in
different parts of the brace-expanded set.  It does also mean that
nounique:true is gone, but I think that's a very acceptable trade-off.