benchmark updates
Only benchmark against fast-glob, we're already faster than globby ;)

Frankly, I don't really see this lib ever getting faster than fast-glob.
Fast-glob sacrifices some correctness for speed, and I'm just not
willing to do that.  (Nothing wrong with that choice, it's just not the
choice I choose.)  In all but the most extreme edge case patterns, the
speed is comparable, and both libraries already do really just the bare
minimum of file system operations, but this one is doing a bit more
extra computation to manage objects whereas fast-glob just kinda yolos
out of a bunch of weird cases.  Hard to say that's not the right call,
"complete coherence with Bash 5" is an arbitrary esthetic choice.

I do think speed of this library *could* get faster than fast-glob, if
the whole Processor/Pattern OOP implementation was scrapped in favor of
just using string/pattern arrays wherever possible, but frankly that was
so much harder to reason about, I don't think it's really worth the
effort. The point in node-glob's favor is that the underlying filesystem
walking lib is a bit more aggressive with caching and more highly tuned
than @nodelib/fs.walk.

Interestingly, the speed delta does *not* seem to come from the costs of
evaluating and testing patterns against filenames (ie, the actual
matching speed) but almost entirely from the fact that node-glob just
has more objects and in some cases does more filesystem operations.
Which means that fast-glob could probably add comparable correctness and
maintain its speed advantage by switching from picomatch/micromatch to
minimatch. I haven't profiled it in detail, which means it's possible
that this assertion is dead wrong, but my reason for thinking so is that
switching out Minimatch for a much more efficient matcher implementation
did not meaningfully impact performance. It improved a little, like 2%,
but not nearly enough to be worth the maintenance overhead, or adding
the work of re-implementing ignores and other features that it lacked.

And, being between 5x and 500x faster than the previous version of
node-glob is probably more than enough for any of its current users, and
puts it well in the range where I can justify pulling glob support back
into rimraf, which is the real goal here.