fix nocase matching on case-sensitive filesystems
If the platform indicates that the filesystem is likely case-sensitive
(ie, not darwin or win32), and nocase:true is set, we can't know if
that's indicating that the filesystem _itself_ is case-insensitive, or
if the user simply wants to perform case-insensitive matching.

In the case where it is in fact likely case-sensitive, and `nocase:true`
is set, do _not_ set `nocaseMagicOnly`, because filesystem operations
will not be run case-insensitively.

Fix: #526