Revert "bring back main/types for ancient outdated node versions"
This reverts commit 7d7b1157f93e439ffe7862307725813d7ff33525.

The problem is that anyone who isn't setting `moduleResolution` such
that exports can be parsed, will be trying to load ESM types as if they
were CommonJS, which is the whole reason for removing top level
main/types in the first place.

There seems to be no way to make this work that is both correct for
contemporary tooling, and compatible with node versions that have been
dead for years. We must choose the future.

Fix: https://github.com/isaacs/node-glob/issues/559