Restore top level main, types
Trying the approach of "type:module, cjs main".

There are scenarios with TS 5.1 where this will result in incorrect
types being loaded, if `module` and `moduleResolution` do not match.

However, this is a smaller hazard than not being able to use the module
at all. And the only node versions which would potentially be confused
by the "type:module" and a cjs main are long out of support (as in, they
can't even load the code, because they also didn't support private class
members), so that's not a relevant consideration.

If this does work and resolve the issue without any negative
consequences arising for a while, I'll make it the default for tshy, so
that it does not require an additional opt-in configuration.

Fix: https://github.com/isaacs/node-glob/issues/557