Add `includeChildMatches: false` option
This does a (best effort) attempt to ignore all *subsequent* matches
that are a descendant of a previous match.

Note the caveat in the docs, though. It will only ignore what it knows
about. There's no post-hoc evaluation to only return the shortest
results in cases where a deeper match is encountered before its
shallower parent.